Move proxy settings to Advanced with progressive disclosure
The Network/proxy settings sat in General as two always-visible input
boxes with help text — heavy visual weight for a low-level setting most
users never touch. Move them to the Advanced pane (alongside HTTP/1.1
Compatibility, same corporate-VPN/proxy audience) under a new Network
subsection, and collapse the inputs behind a single "Configure proxy"
disclosure.

The fields auto-expand when a proxy is already configured or when a
proxy-related settings search matches, so a set value is never hidden.

Renames GeneralNetworkSettingsSection -> AdvancedNetworkSettingsSection
(component + i18n key path across all 5 locales), moves its search entries
into a dedicated advanced-network-search module wired into the Advanced
pane search, and relocates the proxy-draft tests accordingly.
