+++ src/renderer/src/components/settings/advanced-search.ts
@@ -1,9 +1,11 @@
 import type { SettingsSearchEntry } from './settings-search'
+import { getAdvancedNetworkSearchEntries } from './advanced-network-search'
 import { createLocalizedCatalog } from '@/i18n/localized-catalog'
 import { translate } from '@/i18n/i18n'
 import { translateSearchKeyword } from './settings-search-keywords'
 
 export const getAdvancedPaneSearchEntries = createLocalizedCatalog((): SettingsSearchEntry[] => [
+  ...getAdvancedNetworkSearchEntries(),
   {
     title: translate(
       'auto.components.settings.advanced.search.11eea3da72',
