+++ src/renderer/src/i18n/locales/es.json
@@ -4297,7 +4297,9 @@
           "6627e75c92": "Explicar la compatibilidad HTTP/1.1",
           "e9506d3377": "Compatibilidad HTTP/1.1",
           "8b7a8df299": "Soluciones alternativas de bajo nivel para la resolución de problemas de soporte.",
-          "8d8d8ac599": "Compatibilidad"
+          "8d8d8ac599": "Compatibilidad",
+          "network": "Red",
+          "networkDescription": "Enrutamiento de red a nivel de aplicación para proxies y entornos corporativos."
         },
         "AgentLocationSetting": {
           "92f4238f1a": "WSL predeterminado",
@@ -4785,7 +4787,7 @@
           "d21136d9ef": "Configure cómo Orca persiste en las ediciones de archivos.",
           "45c6e85c4d": "Editor"
         },
-        "GeneralNetworkSettingsSection": {
+        "AdvancedNetworkSettingsSection": {
           "3e431564b5": "host local, 127.0.0.1, *.interno",
           "33ee3ca3af": "Opcional. Separe los hosts con comas, punto y coma o líneas nuevas.",
           "f6d76cc8f4": "Reglas de omisión de proxy",
@@ -4796,7 +4798,8 @@
           "f00daf6324": "Proxy HTTP",
           "823e0f15b1": "URL proxy para solicitudes de red Orca y terminales secundarios locales.",
           "d93c7cd531": "Configure el enrutamiento de red a nivel de aplicación.",
-          "c46cdbbd4e": "Red"
+          "c46cdbbd4e": "Red",
+          "configureProxy": "Configurar proxy"
         },
         "GeneralPane": {
           "d58fccfd84": "Navegación",
