+++ src/renderer/src/i18n/locales/ja.json
@@ -4282,7 +4282,9 @@
           "6627e75c92": "HTTP/1.1の互換性について説明する",
           "e9506d3377": "HTTP/1.1 互換性",
           "8b7a8df299": "サポートのトラブルシューティングのための低レベルの回避策。",
-          "8d8d8ac599": "互換性"
+          "8d8d8ac599": "互換性",
+          "network": "ネットワーク",
+          "networkDescription": "プロキシや企業環境向けのアプリレベルのネットワークルーティング。"
         },
         "AgentLocationSetting": {
           "92f4238f1a": "WSL のデフォルト",
@@ -4770,7 +4772,7 @@
           "d21136d9ef": "Orca がファイル編集を保持する方法を構成します。",
           "45c6e85c4d": "エディタ"
         },
-        "GeneralNetworkSettingsSection": {
+        "AdvancedNetworkSettingsSection": {
           "3e431564b5": "ローカルホスト、127.0.0.1、*.internal",
           "33ee3ca3af": "オプション。ホストはカンマ、セミコロン、または改行で区切ります。",
           "f6d76cc8f4": "プロキシ バイパス ルール",
@@ -4781,7 +4783,8 @@
           "f00daf6324": "HTTPプロキシ",
           "823e0f15b1": "Orca ネットワーク リクエストとローカル terminal の子のプロキシ URL。",
           "d93c7cd531": "アプリレベルのネットワークルーティングを構成します。",
-          "c46cdbbd4e": "ネットワーク"
+          "c46cdbbd4e": "ネットワーク",
+          "configureProxy": "プロキシを設定"
         },
         "GeneralPane": {
           "d58fccfd84": "ナビゲーション",
