+++ src/renderer/src/i18n/locales/zh.json
@@ -4282,7 +4282,9 @@
           "6627e75c92": "解释 HTTP/1.1 兼容性",
           "e9506d3377": "HTTP/1.1 兼容性",
           "8b7a8df299": "支持故障排除的低级解决方法。",
-          "8d8d8ac599": "兼容性"
+          "8d8d8ac599": "兼容性",
+          "network": "网络",
+          "networkDescription": "用于代理和企业环境的应用级网络路由。"
         },
         "AgentLocationSetting": {
           "92f4238f1a": "WSL 默认值",
@@ -4770,7 +4772,7 @@
           "d21136d9ef": "配置 Orca 如何保留文件编辑。",
           "45c6e85c4d": "编辑"
         },
-        "GeneralNetworkSettingsSection": {
+        "AdvancedNetworkSettingsSection": {
           "3e431564b5": "本地主机，127.0.0.1，*.internal",
           "33ee3ca3af": "可选。用逗号、分号或换行符分隔主机。",
           "f6d76cc8f4": "代理绕过规则",
@@ -4781,7 +4783,8 @@
           "f00daf6324": "HTTPAgent",
           "823e0f15b1": "Orca 网络请求和本地 terminal 子项的 Agent URL。",
           "d93c7cd531": "配置应用程序级网络路由。",
-          "c46cdbbd4e": "网络"
+          "c46cdbbd4e": "网络",
+          "configureProxy": "配置代理"
         },
         "GeneralPane": {
           "d58fccfd84": "导航",
